Start refactoring the forms into two groups since the forms are getting kind of long.
